machine-learning/requirements.txt
flask
torchvision
pillow
torch